attempt to fix CI, no. 3
